pyproject.toml
[project]
name = "aabp-agent"
version = "1.0.0"
description = "Autonomous Branding Agent — LinkedIn personal branding via OpenClaw + Telegram"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
linkedin-oauth = "skills.linkedin_publish.scripts.linkedin_oauth:main"
linkedin-post = "skills.linkedin_publish.scripts.linkedin_post:main"
